requirements.txt
fastapi
uvicorn
python-dotenv
requests
flask
google-generativeai
pandas
matplotlib

